pyproject.toml
[tool.poetry]
name = "ai-news-intelligence-platform"
version = "0.1.0"
description = ""
authors = ["Ciprian2607 <cmunteanu01@yahoo.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
torch = ">=2.9.1,<3.0.0"
transformers = ">=4.57.3,<5.0.0"
datasets = ">=4.4.1,<5.0.0"
numpy = ">=2.3.5,<3.0.0"
tqdm = ">=4.67.1,<5.0.0"
accelerate = ">=0.30,<1.0"
scipy = "^1.13"
scikit-learn = "^1.5"
pandas = "^2.2"
requests = "^2.31"
feedparser = "^6.0"
beautifulsoup4 = "^4.12"
python-telegram-bot = "^21.0"
schedule = "^1.2"
python-dotenv = "^1.0"
peft = ">=0.11,<1.0"
pyyaml = "^6.0"
black = "^24.0"
ruff = "^0.6"
isort = "^5.13"
mypy = "^1.10"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
